feat: add support to raise custom exceptions on rule validation failures
