+++ docs/rules.rst
@@ -49,3 +49,45 @@ Also you can add all the logic that you need to generate your validations by pas
 
     rule.execute({'age': 11})
     # => True
+
+Rules Custom Exceptions
+-----------------------
+
+Rules can store a custom Exception objects than can be used to be raised when the rule execution is false. This
+custom exception is also used by the `RuleSet` to be raised when the set validation catch a rule failure.
+
+**Stand alone execution**
+
+.. code-block:: python
+
+    from pyruler import Rule
+
+    rule = Rule(name='rule', resolver=lambda x: False, error=AssertionError('custom error'))
+
+    try:
+        if not rule.execute({})
+            raise rule.error
+    except Exception as error:
+        print(error.__class__.__name__, error.args[0])
+        # => AssertionError custom error
+
+**Execution from a RuleSet policy**
+
+.. code-block:: python
+
+    from pyruler import Rule, RuleSet
+
+    rule = Rule(name='rule', resolver=lambda x: False, error=AssertionError('custom error'))
+
+    policy = RuleSet(name='policy')
+    policy.add_rule(rule)
+
+    try:
+        policy.apply({})
+    except Exception as error:
+        print(error.__class__.__name__, error.args[0])
+        # => AssertionError custom error
+
+.. attention::
+    Custom rule errors just will be raised from a RuleSet if the RuleSet policy is running with `fail_fast` flag
+    as True.
