+++ pyproject.toml
@@ -25,7 +25,7 @@ coverage = "~=5.5"
 mamba = "~=0.11.2"
 expects = "~=0.9.0"
 yapf = "~=0.31.0"
-pylint = "~=2.8.2"
+pylint = "~=2.9.3"
 isort = "~=5.9.1"
 radon = "~=5.0.1"
 Sphinx = "^4.0.2"
