+++ pyruler/__init__.py
@@ -1,3 +1,5 @@
-"""Export resources."""
+"""Pyruler main API objects to generate validations."""
 
-from .pyruler import Rule, Ruler, RuleSet
+from .pyruler import Ruler
+from .rule import Rule
+from .ruleset import RuleSet
